+++ R/iorder.R
@@ -25,6 +25,10 @@
 
 
 iorder <- function(obj = NULL, var_name = NULL) {
+  
+  # Deactivate styler cache to avoid blocking message at
+  # first launch
+  styler::cache_deactivate(verbose = FALSE)
 
   run_as_addin <- ifunc_run_as_addin()
 
@@ -42,7 +46,7 @@ iorder <- function(obj = NULL, var_name = NULL) {
     if (is.character(obj) && length(obj) == 1) {
       obj_name <- obj
       try({
-        obj <- get(obj_name, envir = sys.parent())
+        obj <- get(obj_name, envir = .GlobalEnv)
       }, silent = TRUE)
     }
     else {
@@ -54,7 +58,7 @@ iorder <- function(obj = NULL, var_name = NULL) {
       obj_name <- gsub("^\\s*", "", s[[1]][1])
       var_name <- gsub("\\s*$", "", s[[1]][2])
       var_name <- gsub("`", "", var_name)
-      obj <- get(obj_name, envir = sys.parent())      
+      obj <- get(obj_name, envir = .GlobalEnv)      
     }
     if (inherits(obj, "tbl_df") || inherits(obj, "data.table")) obj <- as.data.frame(obj)
 
@@ -117,9 +121,9 @@ iorder <- function(obj = NULL, var_name = NULL) {
                        gettext("Data frame or vector to recode from", domain="R-questionr"),
                        choices = Filter(
                          function(x) {
-                           inherits(get(x, envir = sys.parent()), "data.frame") ||
-                             is.vector(get(x, envir = sys.parent())) ||
-                             is.factor(get(x, envir = sys.parent()))
+                           inherits(get(x, envir = .GlobalEnv), "data.frame") ||
+                             is.vector(get(x, envir = .GlobalEnv)) ||
+                             is.factor(get(x, envir = .GlobalEnv))
                          }, ls(.GlobalEnv)),
                        selected = obj_name, multiple = FALSE)),
               column(6, uiOutput("varInput")))),
@@ -194,7 +198,7 @@ iorder <- function(obj = NULL, var_name = NULL) {
       out <- "<ol id='sortable' class='sortable'>"
       ## List of levels
       if (is.factor(rvar())) levs <- levels(rvar())
-      else levs <- stats::na.omit(unique(rvar()))
+      else levs <- sort(stats::na.omit(unique(rvar())))
       ## Generate fields
       for (l in levs) out <- paste0(out,
                                     '<li><span class="glyphicon glyphicon-move"> </span>&nbsp; <span class="level">',